Move PSFKernel and PSFMap
