+++ gammapy/cube/__init__.py
@@ -4,7 +4,5 @@ from .background import *
 from .exposure import *
 from .fov import *
 from .make import *
-from .psf_kernel import *
-from .psf_map import *
 from .ring import *
 from .simulate import *
