+++ gammapy/cube/make.py
@@ -2,15 +2,64 @@
 import logging
 import numpy as np
 from astropy.coordinates import Angle
-from gammapy.irf import EnergyDependentMultiGaussPSF, EDispMap
+from gammapy.irf import EnergyDependentMultiGaussPSF, EDispMap, PSFMap
 from gammapy.maps import Map
 from .background import make_map_background_irf
 from .exposure import make_map_exposure_true_energy
-from .psf_map import make_psf_map
 
 __all__ = ["MapDatasetMaker", "SafeMaskMaker"]
 
 log = logging.getLogger(__name__)
+
+
+def make_psf_map(psf, pointing, geom, exposure_map=None):
+    """Make a psf map for a single observation
+
+    Expected axes : rad and true energy in this specific order
+    The name of the rad MapAxis is expected to be 'rad'
+
+    Parameters
+    ----------
+    psf : `~gammapy.irf.PSF3D`
+        the PSF IRF
+    pointing : `~astropy.coordinates.SkyCoord`
+        the pointing direction
+    geom : `~gammapy.maps.Geom`
+        the map geom to be used. It provides the target geometry.
+        rad and true energy axes should be given in this specific order.
+    exposure_map : `~gammapy.maps.Map`, optional
+        the associated exposure map.
+        default is None
+
+    Returns
+    -------
+    psfmap : `~gammapy.cube.PSFMap`
+        the resulting PSF map
+    """
+    energy_axis = geom.get_axis_by_name("energy_true")
+    energy = energy_axis.center
+
+    rad_axis = geom.get_axis_by_name("theta")
+    rad = rad_axis.center
+
+    # Compute separations with pointing position
+    offset = geom.separation(pointing)
+
+    # Compute PSF values
+    # TODO: allow broadcasting in PSF3D.evaluate()
+    psf_values = psf._interpolate(
+        (
+            rad[:, np.newaxis, np.newaxis],
+            offset,
+            energy[:, np.newaxis, np.newaxis, np.newaxis],
+        )
+    )
+
+    # TODO: this probably does not ensure that probability is properly normalized in the PSFMap
+    # Create Map and fill relevant entries
+    data = psf_values.to_value("sr-1")
+    psfmap = Map.from_geom(geom, data=data, unit="sr-1")
+    return PSFMap(psfmap, exposure_map)
 
 
 def make_edisp_map(edisp, pointing, geom, exposure_map=None):
