+++ gammapy/datasets/map.py
@@ -8,8 +8,8 @@ from astropy.table import Table
 from astropy.utils import lazyproperty
 from regions import CircleSkyRegion
 from gammapy.irf.edisp_map import EDispMap
-from gammapy.cube.psf_kernel import PSFKernel
-from gammapy.cube.psf_map import PSFMap
+from gammapy.irf.psf_kernel import PSFKernel
+from gammapy.irf.psf_map import PSFMap
 from gammapy.data import GTI
 from gammapy.irf import EDispKernel, EffectiveAreaTable
 from gammapy.maps import Map, MapAxis
