+++ gammapy/irf/__init__.py
@@ -10,6 +10,8 @@ from .io import *
 from .irf_reduce import *
 from .irf_stack import *
 from .psf_3d import *
+from .psf_map import *
 from .psf_gauss import *
 from .psf_king import *
 from .psf_table import *
+from .psf_kernel import *
