+++ gammapy/irf/psf_kernel.py
@@ -3,9 +3,10 @@ import numpy as np
 import astropy.units as u
 from astropy.coordinates import Angle
 from astropy.coordinates.angle_utilities import angular_separation
-from gammapy.irf import TablePSF
 from gammapy.maps import Map, WcsGeom
 from gammapy.utils.gauss import Gauss2DPDF
+from .psf_table import TablePSF
+
 
 __all__ = ["PSFKernel"]
